Empty block test using AST
