+++ examples/empty_block.yul
@@ -1,0 +1,1 @@
+{ }
