+++ examples/power_function_signature.yul
