+++ src/lib.rs
@@ -1,2 +1,7 @@
+extern crate pest;
+#[macro_use]
+extern crate pest_derive;
+
 pub mod yul;
 pub mod validator;
+pub mod yul_parser;
